labAgent/requirements.txt
uagents
fastmcp
httpx
python-dotenv
openai
psycopg2-binary
sqlalchemy
smtplib
email-validator